Finishes TF-IDF computing

+++ src/cs480a2/yqiu/recSystem/mapreduce/input/SingleBookReader.java
@@ -133,9 +133,9 @@ public class SingleBookReader extends RecordReader<Text, Text> {
         currentPos += readBytes;
 
         if (currentLine.toString().startsWith("End of Project Gutenberg")) {
-            int totalCount = configuration.getInt("Total_Book_Count", 0);
+            double totalCount = configuration.getDouble("Total_Book_Count", 0);
             totalCount++;
-            configuration.setInt("Total_Book_Count", totalCount);
+            configuration.setDouble("Total_Book_Count", totalCount);
             return false;
         }
